delete entry with offset in same arg or not

+++ includes/history.h
@@ -17,7 +17,7 @@ typedef	struct s_hist_opt
 	bool	w;
 	bool	p;
 	bool	s;
-	size_t	offset;
+	int		offset;
 	char	*filename;
 	char	**args;
 }				t_hist_opt;
